basic abstracion build, need refactor and correction of tests
